requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-api-response==0.1.0
django-redis==5.4.0
django-environ==0.9.0
django-ckeditor==6.3.2

pillow==11.0.0

uvicorn==0.32.0

channels==4.1.0
channels-redis==4.2.0

whitenoise==6.8.2

psycopg2==2.9.4


celery==5.4.0
django-celery-results==2.5.1
django-celery-beat==2.7.0

django-storages==1.14.4
boto3==1.35.67
botocore==1.35.67
cryptography==41.0.7
rsa==4.9